backend/requirements.txt
asgiref==3.6.0
Django==4.2
django-appconf==1.0.5
django-cors-headers==3.14.0
django-imagekit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
pilkit==2.0
Pillow==9.4.0
pytz==2023.2
six==1.16.0
sqlparse==0.4.3
